: Add scripts for browser automation with Proxies
A script for web scraping with proxies has been implemented using Python 3.9, Playwright, Tkinter, and Asyncio. This script allows the user to provide a list of URLs and a list of proxies, in which the script will automate visiting each URL using a different proxy. Tkinter was used to create a user interface for better user experience while Playwright was used to control the web browser actions. Asyncio was implemented for better handling of multiple webpage loads at the same time.

This addition will allow users to perform web scraping tasks on multiple websites simultaneously with proxies for better efficiency and data collection. This can simplify web scraping tasks that require visiting multiple URLs while maintaining a level of anonymity provided by the use of proxies.